chore: create separate AutoplayControls component
