+++ src/pages/view/slideshow/styles/SlideshowControlbar.module.css
@@ -28,6 +28,7 @@
 
 .end {
     justify-content: end;
+    gap: 4px;
 }
 
 .button {
